requirements.txt
streamlit
streamlit-navigation-bar
pandas
numpy
altair
matplotlib
streamlit_code_editor